Updated GitHub references and defined language env variables
Updated the GitHub reference for SATOSA to use PR 243 until
it is merged. Also update the reference for pysaml2 to a specific
commit now that the PR for enhanced eduPersonTargetedID functionality
has been merged. Defined LC_ALL and LANG environment variables so
that it is easier to break into the container and run
satosa-saml-metadata.